requirements.txt
pytest==7.3.1
webdriver-manager==4.0.1
pytest-html==3.2.0
pytest-xdist==3.3.1
openpyxl==3.1.2
allure-pytest==2.13.2
selenium==4.9.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
contextlib2==21.6.0
distlib==0.3.7
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
execnet==2.0.2
filelock==3.12.4
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mock==5.1.0
MouseInfo==0.1.3
mss==9.0.1
openpyxl==3.1.2
outcome==1.2.0
packaging==23.2
path==16.7.1
path.py==12.5.0
pillow==10.0.0
platformdirs==3.5.1
pluggy==1.3.0
py==1.11.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
pyRect==0.2.0
pyscreenshot==3.1
pyscreeze==0.1.29
PySocks==1.7.1
pytest==7.3.1
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-shutil==1.7.0
pytest-xdist==3.3.1
python-dotenv==1.0.0
pytweening==1.0.7
requests==2.31.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
termcolor==2.3.0
tomli==2.0.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
urllib3==1.26.18
wsproto==1.2.0
XlsxWriter==3.1.9

